id representation experiments

+++ src/lib/models/decode.py
@@ -87,10 +87,10 @@ def mot_decode(heat, wh, reg=None, ltrb=False, K=100, id_feature=None):
         box_indx2, box_indy2 = xs + wh[..., 2:3] / 2, ys + wh[..., 3:4] / 2
     detections = torch.cat([bboxes, scores, clses], dim=2)
 
-    ind_lt = (torch.mul(box_indy1, width) + box_indx1).int().squeeze(2)
-    ind_rt = (torch.mul(box_indy1, width) + box_indx2).int().squeeze(2)
-    ind_lb = (torch.mul(box_indy2, width) + box_indx1).int().squeeze(2)
-    ind_rb = (torch.mul(box_indy2, width) + box_indx2).int().squeeze(2)
+    ind_lt = (torch.mul(box_indy1, width) + box_indx1).to(torch.int64).squeeze(2)
+    ind_rt = (torch.mul(box_indy1, width) + box_indx2).to(torch.int64).squeeze(2)
+    ind_lb = (torch.mul(box_indy2, width) + box_indx1).to(torch.int64).squeeze(2)
+    ind_rb = (torch.mul(box_indy2, width) + box_indx2).to(torch.int64).squeeze(2)
 
     # id_feature = _tranpose_and_gather_feat(id_feature, inds)
     # 方案一：相加
